services/scheduler/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.25.1
aiohttp==3.9.1

# Optimization Libraries
pulp==2.7.0
numpy==1.24.3
scipy==1.11.4
ortools==9.7.2996
networkx==3.2.1

# PDF Generation
reportlab==4.0.7
pillow==10.1.0

# Authentication
pyjwt==2.8.0
passlib==1.7.4
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3
python-jose[cryptography]==3.3.0

# Logging
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0